Fix rounding errors of mrsteveman1's stage 3 votes
They added up to 1001 instead of the allowed 1000.
This was because he exported with rounding, which converted the default
value of 14.925 to 15 for tasks for which he didn't change the default.

I've asked him on IRC and he was OK with me changing the 15 to 14.925
everywhere.
As a conclusion, "results-raw-in-one-sheet" is renamed to
"results-in-one-sheet".